cloudleak_backend/requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.79.0
Flask==2.2.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
loguru==0.6.0
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pydantic==1.9.1
pymongo==4.2.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
